Style: 리코일 코드 정리
